Add github clone option
