+++ electron-app/src/python/pyrunner.py
@@ -395,6 +395,13 @@ def post(request):
                 ),
                 "returncode": 0,
             }
+        elif query == "settings/github-clone":
+            create_folder = data.get("createfolder", True);
+            data = {
+                "query": query,
+                "body": runner.github.Clone(data["url"], data["path"], create_folder),
+                "returncode": 0,
+            }
 
         else:
             raise NotImplementedError(f"Unknown query: {query}")
